Reworked Assert.expectedException() to be more JavaScripty
Assert.expectedException() now wraps a function so that it can validate
the exception inline the same as any of the other assert functions. This
cleans up the _execTest() code as it now no longer needs to check the
exceptions it catches to see if we were expecting it or not.